+++ test/test_generated_mypy.py
@@ -33,7 +33,12 @@ from testproto.test_extensions3_pb2 import (
     repeated_scalar_option,
     scalar_option,
 )
-from testproto.test_pb2 import DESCRIPTOR, FOO, Extensions1, Extensions2
+from testproto.test_pb2 import (
+    DESCRIPTOR,
+    FOO,
+    Extensions1,
+    Extensions2,
+)
 from testproto.test_pb2 import Name as NamingConflicts_Name
 from testproto.test_pb2 import (
     NamingConflicts,
@@ -63,10 +68,10 @@ def _is_summary(line: str) -> bool:
 
 def test_generate_mypy_matches() -> None:
     proto_files = glob.glob("proto/**/*.proto", recursive=True)
-    assert len(proto_files) == 17  # Just a sanity check that all the files show up
+    assert len(proto_files) == 18  # Just a sanity check that all the files show up
 
     pyi_files = glob.glob("test/generated/**/*.pyi", recursive=True)
-    assert len(pyi_files) == 19  # Should be higher - because grpc files generate extra pyis
+    assert len(pyi_files) == 20  # Should be higher - because grpc files generate extra pyis
 
     failure_check_results = []
     for fn in proto_files:
@@ -99,6 +104,9 @@ def test_generate_negative_matches() -> None:
             if _is_summary(line):
                 continue
             parts = line.split(":")
+            # Filter out errors not from negative.py. Like deprecation warnigns from pyi files that were import followed
+            if not parts[0].endswith("test_negative/negative.py"):
+                continue
             yield parts[0], int(parts[1])
 
     def grab_expectations(filename: str, marker: str) -> Generator[Tuple[str, int], None, None]:
@@ -106,14 +114,14 @@ def test_generate_negative_matches() -> None:
             if "#" in line and marker in line:
                 yield filename, idx + 1
 
-    errors_38 = set(grab_errors("test_negative/output.expected.3.8"))
+    errors_39 = set(grab_errors("test_negative/output.expected.3.9"))
 
-    expected_errors_38 = set(grab_expectations("test_negative/negative.py", "E:3.8"))
+    expected_errors_39 = set(grab_expectations("test_negative/negative.py", "E:3.8"))
 
-    assert errors_38 == expected_errors_38
+    assert errors_39 == expected_errors_39
 
     # Some sanity checks to make sure we don't mess this up. Please update as necessary.
-    assert len(errors_38) == 77
+    assert len(errors_39) == 84
 
 
 def test_func() -> None:
@@ -516,3 +524,30 @@ def test_reserved_keywords() -> None:
     prk = PythonReservedKeywords(none=none_instance, valid=PythonReservedKeywords.valid_in_finally)
     assert prk.none.valid == 5
     assert prk.valid == PythonReservedKeywords.valid_in_finally
+
+
+def test_editions_2024() -> None:
+    from testproto.edition2024_pb2 import Editions2024SubMessage, Editions2024Test
+
+    submsg = Editions2024SubMessage(thing="example")
+
+    testmsg = Editions2024Test(
+        legacy="legacy value",
+        explicit_singular="explicit value",
+        message_field=submsg,
+        implicit_singular="implicit value",
+        default_singular="default value",
+    )
+    assert testmsg.legacy == "legacy value"
+    assert testmsg.explicit_singular == "explicit value"
+    assert testmsg.message_field == submsg
+    assert testmsg.implicit_singular == "implicit value"
+    assert testmsg.default_singular == "default value"
+
+    assert testmsg.HasField("explicit_singular")
+    assert testmsg.HasField("message_field")
+    assert testmsg.HasField("legacy")
+    assert testmsg.HasField("default_singular")
+
+    with pytest.raises(ValueError):
+        testmsg.HasField("implicit_singular")  # type: ignore